frontend/requirements.txt
Flask==2.2.2
google-cloud-dialogflow==2.19.1
gunicorn==20.1.0